+++ frontend/src/constans/constans.ts
@@ -1,0 +1,2 @@
+export const URL = "http://localhost:3007/api/city";
+export const LIMIT = 15;
